+++ snuba_sdk/entity.py
@@ -1,8 +1,9 @@
 import re
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from typing import Optional
 
 from snuba_sdk.expressions import Expression
+from snuba_sdk.schema import EntityModel
 
 entity_name_re = re.compile(r"^[a-zA-Z_]+$")
 
@@ -11,11 +12,12 @@ class InvalidEntityError(Exception):
     pass
 
 
-@dataclass(frozen=True)
+@dataclass(frozen=True, repr=False)
 class Entity(Expression):
     name: str
     alias: Optional[str] = None
     sample: Optional[float] = None
+    data_model: Optional[EntityModel] = field(hash=False, default=None)
 
     def validate(self) -> None:
         # TODO: There should be a whitelist of entity names at some point
@@ -31,6 +33,17 @@ class Entity(Expression):
         if self.alias is not None:
             if not isinstance(self.alias, str) or not self.alias:
                 raise InvalidEntityError(f"'{self.alias}' is not a valid alias")
+
+        if self.data_model is not None:
+            if not isinstance(self.data_model, EntityModel):
+                raise InvalidEntityError(
+                    "data_model must be an instance of EntityModel"
+                )
+
+    def __repr__(self) -> str:
+        alias = f", alias='{self.alias}'" if self.alias is not None else ""
+        sample = f", sample={self.sample}" if self.sample is not None else ""
+        return f"Entity('{self.name}'{alias}{sample})"
 
 
 # TODO: This should be handled by the users of the SDK, not the SDK itself.